Move Project sidebar panels to seperate file
